+++ CHANGELOG.rst
@@ -1,6 +1,63 @@
 Changelog
 ---------
 
+0.4.6 (2026-08-14)
+******************
+
+New:
+
+- Add GeomAI prediction dimension reduction support.
+
+Changes:
+
+- Standardize the :py:meth:`optimizations.get<ansys.simai.core.data.optimizations.OptimizationDirectory.get>` method to use `id` as argument instead of `optimization_id`. `optimization_id` is now deprecated.
+- Added an `optimization` property to the :py:meth:`Optimization<ansys.simai.core.data.optimizations.Optimization>` class to match API with `LegacyOptimization`. This will be removed in the future, along with `LegacyOptimization`.
+
+Fixes:
+
+- Use correct ``client_id`` when using server side optimization based on model's version.
+
+Documentation:
+
+- Refactor resolution section in ``Generative Design`` documentation.
+- Update PySimAI upload examples.
+
+0.4.5 (2026-07-28)
+******************
+
+New:
+
+- Add GeomAI prediction auto-resolution.
+
+Changes:
+
+- Update timeout value and chunk size for file upload. This should improve upload stability.
+
+Fixes:
+
+- Handle remote protocol error in SSE stream.
+
+Documentation:
+
+- Update GeomAI resolution documentation.
+
+0.4.4 (2026-06-19)
+******************
+
+Changes:
+
+- Deprecate ``download_latent_parameters_json`` method from GeomAI workspace.
+
+Fixes:
+
+- Fix API endpoint called when retrieving GeomAI workspace model configuration.
+- Preserve legacy ``client_id`` when refreshing tokens issued with it.
+
+Documentation:
+
+- Fix compatibility issue with new optimization object.
+- Add new example for bulk upload.
+
 0.4.3 (2026-05-19)
 ******************